feat(harness): json output support
The overall design is "re-implement all other output on top of json".
This allows us to dog-food the format.  In particular, I want to
eventually move all of this up into `cargo test` for when we do parallel
binary runs.

Building on top of json output allows a decoupling of running and
reporting that makes it easier to support threading.

Warning: this deviates from the official json format

Fixes #10