feat(harness): Allow tests to be declared as `exclusive`
`libtest` runs benchmarks in an exclusive mode in a pattern similar  to
what the current state of the harness.  Instead of keying off of being a
`TestKind::Bench`, we are giving control to the `Case` to determine what
to do which will allow deterministic benchmarks to run concurrently and
allow (badly behaved) tests to run exclusively.